Reduce `batch-open-conversations` to `open-all-conversations`
